refactor : 이벤트 코드 useProject 관리
